move block stream out of sequencer module
